Substitute the pico header with dummy type definition in the apistub.cpp.
Previous code has redefinition error during the Pico compile.
The root cause of this error is inline funciton definition inside pico header.
To avoid the error, we should not use the pico header for stub.

The dummy type for the stub is OK, because we don't use any
parameters inside stub.