Wrap entire utility classed by __has_include.
This is very complicated. Because of this class library is a collection,
entire library will be build even  some are not used. In the cae certain
hardwaer/xxx library in the Pico SDK, depending class in this collection
will be fail to compile.

So, we need to wrap the utility classes by __has_include to avoid this
problem.